Moved GraphMetadata calculation from runtime to buildTime
Now GraphMetadata is calculated after OSM is loaded.
This info is used during linking phase to skip stops which aren't in
OSM envelope.
During GTFS building each stop is added to GraphMetadata to
expandEnvelope and TransitMode of Stop is added.

After all stops are linked envelope with GTFS and OSM is used as
primary.