Added travel options to RouterInfo
This is a list of "Travel by" options which can be seen in dropdown in
a client. It is created based on supported transit modes and extra
options (B & R, P & R, bike sharing).
It consists of values to send to server if it is chosen and name.

The client needs to localize name using prefered locale framework.