Add tests for permissions in street edges with bollard
Since there are two kinds of permissions:
- canTraverse(TraverseModeSet) which just checks street permissions
 without accounting  for bollard
- canTraverse(RoutingRequest, TraverseMode) which is used in doTraverse
  and uses permissions of bollard

First is used in searching for searching for nearest edges and car/walk
edges in Park and ride.
Second is used in actuall traversing. So it can happen that edge has
canTraverse true but actually can't be traversed with specific mode.

Edge for example has bollard at the end and isn't traversable with a
car, but destination is couple of meters before the bollard and when
edge is split for that, edge is traversable also for CAR.

Fixes #1914