Store edge geometries, share info between fore/back edge pairs
The serialized graphs now contain complete geometry information, yet are somewhat smaller than before because a lot of information is now being shared between the two edges in a pair. Loading and saving are up to twice as slow, probably because of all the array objects holding the geometries and the lists of pointers to them. Perhaps they should be replaced with a single contiguous array of delta-compacted points.