Refactor the object differ for use in tests. #2643.
This enables reliable tests that serialization / deserialization perfectly restores a Graph.
In some ways this simplifies the differ, but it also ensures that it performs a very deep recursive comparison of almost every object.
Also added tests that can compare two references to the same object, which provide strong evidence that the differ actually works.